Fix bug letting two RE elements run
When compressor is lockout and has backup pointing to lower element only run if the VIP element isn't running. Happens during a large draw even when air temperature is below compressor lockout.

Rheem2020 tests change because the tanks are on the smaller end of the tested spectrum and have an upper element with the compressor as a companion heat source